Restructuring bigquery and spark consumers (#73)
* Log metric values as info message when MaxBatchSizeWait is used

* Restructuring bigquery and spark consumers

(cherry picked from commit 1e158f728519f96ace56dc4060bfe34b4a614451)

+++ debezium-server-batch-sinks/src/main/java/io/debezium/server/batch/common/InterfaceBatchSizeWait.java
@@ -19,6 +19,6 @@ public interface InterfaceBatchSizeWait {
   default void initizalize() {
   }
 
-  void waitMs(Integer numRecordsProcessed, Integer processingTimeMs) throws InterruptedException;
+  void waitMs(long numRecordsProcessed, Integer processingTimeMs) throws InterruptedException;
 
 }
